# src/main/java/cool/cfapps/reactivelinkappxplatform/dto/CategoryDto.java
package cool.cfapps.reactivelinkappxplatform.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class CategoryDto {
    private long id;
    private String title;
}