# src/main/java/cool/cfapps/reactivelinkappxplatform/dto/CreateCategoryDto.java
package cool.cfapps.reactivelinkappxplatform.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class CreateCategoryDto {
    private String title;
}